+++ tests/testthat/test-set_math.R
@@ -1,0 +1,130 @@
+test_that("math operations", {
+  options(cheapr.cores = 2)
+
+  # The first 2 will not trigger multi-threaded calculations
+  make_test_data1 <- function(){
+    set.seed(3742)
+    assign("x", fill_with_na(rnorm(10^3), 50),
+           envir = parent.frame())
+  }
+  make_test_data2 <- function(){
+    set.seed(3742)
+    assign("x", fill_with_na(sample.int(100, 10^3, TRUE), 50),
+           envir = parent.frame())
+  }
+  make_test_data3 <- function(){
+    set.seed(3742)
+    assign("x", fill_with_na(rnorm(10^5), 10^3),
+           envir = parent.frame())
+  }
+  make_test_data4 <- function(){
+    set.seed(3742)
+    assign("x", fill_with_na(sample.int(100, 10^5, TRUE), 10^3),
+           envir = parent.frame())
+  }
+
+  expect_error(set_abs(iris))
+  expect_error(set_floor(iris))
+  expect_error(set_ceiling(iris))
+  expect_error(set_trunc(iris))
+  expect_error(set_log(iris))
+  expect_error(set_exp(iris))
+  expect_error(set_sqrt(iris))
+  expect_error(set_round(iris))
+
+  make_test_data1()
+  expect_equal(
+   round(x, 2),
+   set_round(x, 2)
+  )
+  make_test_data1()
+  expect_equal(
+    round(x),
+    set_round(x)
+  )
+  make_test_data1()
+  expect_equal(round(x, 1),set_round(x, 1))
+  make_test_data1()
+  expect_equal(abs(x), set_abs(x))
+  make_test_data1()
+  expect_identical(floor(x), set_floor(x))
+  make_test_data1()
+  expect_identical(ceiling(x), set_ceiling(x))
+  make_test_data1()
+  expect_identical(trunc(x), set_trunc(x))
+  make_test_data1()
+  expect_equal(exp(x), set_exp(x))
+  make_test_data1()
+  expect_equal(log(abs(x)), set_log(set_abs(x)))
+  make_test_data1()
+  expect_equal(log10(abs(x)), set_log(set_abs(x), base = 10))
+  make_test_data1()
+  expect_equal(sqrt(abs(x)), set_sqrt(set_abs(x)))
+  make_test_data1()
+  expect_equal(-x, set_change_sign(x))
+
+
+  make_test_data2()
+  expect_equal(round(x, 1),set_round(x, 1))
+  make_test_data2()
+  expect_equal(abs(x), set_abs(x))
+  make_test_data2()
+  expect_equal(floor(x), set_floor(x))
+  make_test_data2()
+  expect_equal(ceiling(x), set_ceiling(x))
+  make_test_data2()
+  expect_equal(trunc(x), set_trunc(x))
+  make_test_data2()
+  expect_equal(exp(x), set_exp(x))
+  make_test_data2()
+  expect_equal(log(abs(x)), set_log(set_abs(x)))
+  make_test_data2()
+  expect_equal(log10(abs(x)), set_log(set_abs(x), base = 10))
+  make_test_data2()
+  expect_equal(sqrt(abs(x)), set_sqrt(set_abs(x)))
+  make_test_data2()
+  expect_equal(-x, set_change_sign(x))
+
+  make_test_data3()
+  expect_equal(round(x, 1),set_round(x, 1))
+  make_test_data3()
+  expect_equal(abs(x), set_abs(x))
+  make_test_data3()
+  expect_equal(floor(x), set_floor(x))
+  make_test_data3()
+  expect_equal(ceiling(x), set_ceiling(x))
+  make_test_data3()
+  expect_equal(trunc(x), set_trunc(x))
+  make_test_data3()
+  expect_equal(exp(x), set_exp(x))
+  make_test_data3()
+  expect_equal(log(abs(x)), set_log(set_abs(x)))
+  make_test_data3()
+  expect_equal(log10(abs(x)), set_log(set_abs(x), base = 10))
+  make_test_data3()
+  expect_equal(sqrt(abs(x)), set_sqrt(set_abs(x)))
+  make_test_data3()
+  expect_equal(-x, set_change_sign(x))
+
+  make_test_data4()
+  expect_equal(round(x, 1),set_round(x, 1))
+  make_test_data4()
+  expect_equal(abs(x), set_abs(x))
+  make_test_data4()
+  expect_equal(floor(x), set_floor(x))
+  make_test_data4()
+  expect_equal(ceiling(x), set_ceiling(x))
+  make_test_data4()
+  expect_equal(trunc(x), set_trunc(x))
+  make_test_data4()
+  expect_equal(exp(x), set_exp(x))
+  make_test_data4()
+  expect_equal(log(abs(x)), set_log(set_abs(x)))
+  make_test_data4()
+  expect_equal(log10(abs(x)), set_log(set_abs(x), base = 10))
+  make_test_data4()
+  expect_equal(sqrt(abs(x)), set_sqrt(set_abs(x)))
+  make_test_data4()
+  expect_equal(-x, set_change_sign(x))
+  options(cheapr.cores = 1)
+})